feat(types): Add type hints and code optimizations
Python3.8 is almost end-of-life and backwards compatibility
to versions before type hints is not a priority anymore.

This commit is a start and by no means complete or correct
for all instances.

Furthermore I included code optimizations. These mainly regard
raising of exceptions, pytest style tests and some minor
refactoring.